Fix print_markers(), used in debug_cursor().
The end value of 0 to mean the end of the array was used only in one
place, but there are cases where 0 is the actual end of the segment to
print, which would need special handling in the debug_cursor().
It’s much simpler to avoid this convenience special value and call
print_markers() with `markers->len` instead of `0` when needed.