Backstitch: add syntax for composite prefixes with ellipsis.
Now the ellipsis needs to be adjacent to the identifier, for both prefix
and suffix, while before there could be spaces in between.
This allows another case which is ellipsis without prefix or suffix to
indicate that the object part, which can be a composite expression, is
to be used as prefix.