Avoid shadowing #error directives.
When there is an error in a macro and an #error directive is inserted,
that can cause another error in another macro applied later, which would
cause the new error to be inserted just before the previous one and the
file being truncated after the new #error directive.
This caused miselading error reports.