Detect unterminated strings/characters.
These changes include fixing some off-by-one errors that did not have
any consequence because of the source array padding.
Also streamlined some of the error reporting in parse() so that it is
easy to mirror it in parse_skip_until_cedro_pragma(), so that the same
errors are detected and reported in the same way both before and after
the pragma.