Improve error reporting.
The previous method was causing spurious error messages that masked the
real ones.
Now when there is an error, the whole contents is replaced with the
single #error preprocessor directive so there is no confusion.
I thought it would be useful to have the code up to that point, but it
turns out to be unnecessary because we get the line number at the
original file.